several unrelated commits. sorry about that. i was offline and unable to commit while developing.
defer sending of notification emails until cron time. this speeds up posting for big groups with lots of email subscribers. the job_queue module is required for this feature. if not present, notifications are sent in real time (just like before this change).

remove 'website' as a field for groups. in order to get this back, add 'website' to your node type(s) that behave as groups. i left the website field in the og table. sites that want to preserve this data need to migrate it to their new field.

added an og_requirements() which gently warns if og access control is disabled and if job_queue.module is disabled.